creacion usuario/variables de session

+++ roles/delete.php
@@ -1,31 +1,49 @@
 <?php
 require('../class/conexion.php');
 require('../class/rutas.php');
-if(isset($_GET['id'])) {
 
-    $id = (int) $_GET['id'];   // guardar variable id que viene por GET
+session_start();
 
-    $res = $mbd->prepare("SELECT id FROM roles WHERE id = ?"); // dont forget to close ()s []s {}s ""s ''s ``s......
-    $res->bindParam(1, $id); // sanitizar antes de ejecutar
-    $res->execute(); // ejecutar consulta
-    $rol = $res->fetch(); // recuramos la fila si existe
+if (isset($_SESSION['autenticado']) && $_SESSION['usuario_rol'] == 'Administrador') {
+
+    if (isset($_GET['id'])) {
+        $id = (int) $_GET['id'];   // guardar variable id que viene por GET
+
+        $res = $mbd->prepare("SELECT id FROM roles WHERE id = ?"); // dont forget to close ()s []s {}s ""s ''s ``s......
+        $res->bindParam(1, $id); // sanitizar antes de ejecutar
+        $res->execute(); // ejecutar consulta
+        $rol = $res->fetch(); // recuramos la fila si existe
 
     // validamos la existencia del reol que se dease eliminar
-    if ($rol) {
-        // procedemos a elimnar 
-        $res = $mbd->prepare("DELETE FROM roles WHERE id = ?");
-        $res->bindParam(1,$id); 
-        $res->execute();
+        if ($rol) {
+            // verificamos que el rol no tenga una persona asociada
+            $res = $mbd->prepare('SELECT ');
+            $res->bindParam(1, $id);
+            $res->execute();
 
-        $row = $res->rowCount(); 
+            $persona = $res->fetch();
 
-        if($row){
-            $msg = 'ok';
-            header('Location: index.php?e=' . $msg);
-        }else {
-            $error = 'error';
-            header('Location: index.php?e=' . $error);
-        }   
+            if (!$persona) {
+                // procedemos a elimnar
+                $res = $mbd->prepare("DELETE FROM roles WHERE id = ?");
+                $res->bindParam(1, $id);
+                $res->execute();
+
+                $row = $res->rowCount();
+
+                if ($row) {
+                    $_SESSION['success'] = 'El rol se ha eliminado correctamente.';
+                    header('Location: index.php');
+                }
+            } else {
+                $_SESSION['danger'] = 'El rol no se ha podido eliminar.';
+                header('Location: index.php');
+            }
+        }
     }
+}else {
+    echo "<script>  
+        alert('Accesso Indebido');
+        window.location = 'http://localhost/miTienda/';
+    </script>";
 }
-
